refactor(file-service): extract the shared upload helpers
`put` and `validateAndNormalizeFilePathOrThrow` live on the DatasetResource
companion but hold no dataset state: one PUTs bytes to a presigned URL, the
other normalizes a relative path. Model upload needs both, so move them to a
new ResourceUploadUtils rather than copy them.

Pure move: the bodies are unchanged, DatasetResource imports them instead of
defining them, and DatasetResourcePathSpec now exercises them at their new
home. No behavior change.

Part of #6494, groundwork for #6498.
